backend/requirements.txt
# backend/requirements.txt
python-multipart
pip==26.0.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
requests==2.32.5
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
qrcode==8.2
httpx==0.28.1
google-generativeai==0.8.6
tiktoken==0.12.0

